+++ CHANGELOG.md
@@ -80,6 +80,11 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ### Internal improvements
 
+#### exonum
+
+- `system/v1/shutdown` endpoint has been modified and now accepts empty POST
+  requests. (#1526)
+
 #### exonum-merkledb
 
 - `ProofListIndex::extend()` method has been refactored, leading to up to 10x